requirements.txt
python-telegram-bot==20.0a4
psycopg2==2.9.5
